Track RunningPlan in beh, plan's exec context + optimize (#174)
The running plan under which the behaviour or plan is executing was not
tracked (execution context). Instead only the running plan under which
the behaviour or plan was started i.e. signal context was tracked. This
meant that it was possible for the behaviour to access a destroyed
running plan object.

Optimization: Skip calling init/terminate when the behaviour/plan's
execution is out of context with the signal from the main thread.

Improve comments