Fix alica engine termination
Do the following in order to correctly terminate the engine:
1. Stop the plan base thread. This prevents any more changes to the
running plan.
2. Deactivate the running plan in plan base stop(). This ensures the
plans/behaviours are scheduled for termination in the correct order.
3. Stop all behaviours & plans that may be running. There should not
be any since (2) ensures this taken care of
4. Execute all pending non repeatable jobs in the scheduler thread.
This ensures the plan/behaviour init/terminates are actually executed.
5. Stop the scheduler thread

Add a trace log that indicates if the trace is for a behaviour or a
plan.

Note: the master plan trace is not reported because the master plan does
not have a corresponding basic plan attached to it. This is a bug &
the reason for it is because there is no ConfAbstractPlanWrapper for the
master plan due to which it is not in the plan pool. This has to be fixed
in a separate PR.