Use context from root to leaf for success semantics
Success semantics should include the entire context from the root
to the leaf of the running plan instead of just the entry point
information. The context of a running plan is the hash of <dynamic
entry point id, state id, parent's hash>. If a running plan is
successful the context of success is the hash of <parent running plan's
context, entry point id, dynamic entry point id>. This commint includes
the changes to only set the success. Reading the success information
from other agents & subsequent changes to the success data structures
is not included.