Map blackboard output after termination
The doTerminate method cancels the run job in the scheduler followed
by calling the termination method of the application. Since the output
key mapping is done before doTerminate without locking the blackboard,
the keys can be mapped at the same time the run & termination methods
access them leading to race conditions. To fix this simply map the keys
after doTerminate.