Fix scheduler thread wakeups
It was possible for the scheduler thread to wait on its condition
variable while the condition variable was notified during both
termination & addition of a job to the job queue due to race conditions
resulting from the unlocked manipulation of the running flag & the job
queue. Fix this by grabbing the wait mutex before modifying said
variables.