Consider only the file name in referenced elements
If the referenced element has a path instead of a name only consider the
name Eg. if a behaviour is referenced by a path (relative or absolute)
to the .beh file then only consider the file name since we perform
a search for the file now instead of directly using the path.