Add ability to find config files in multiple dirs (#429)
* Add ability to find config files in multiple dirs

- The engine now has a set of directories where it looks for config
files recursively
- If the roleset with the given file name is not found (recursive
search), then the engine looks for any roleset file in the specified
directories but not recursively
- The beh & cnd files are now referred to only by their file name &
not by their relative path to the plans folder
- The plan, role & task config keys are no longer required since the
search is recursive & hence these are removed

* Accept multiple config folders in the AlicaContext

* Consider only the file name in referenced elements

If the referenced element has a path instead of a name only consider the
name Eg. if a behaviour is referenced by a path (relative or absolute)
to the .beh file then only consider the file name since we perform
a search for the file now instead of directly using the path.

* Fix tests

* Fix turtlesim

* Review comments: Change unordered_set to vector

For config paths container, use vector instead of unordered_set so that
user order can be maintained. In this case, the search for files will
proceed from index 0 which means that will also be the order of priority
to find files.
Couple of other very minor fixes

* Add multi config folder plan parsing unit test

* Fix comment

* Remove unnecessary include

* Minor fix

* More fixes

* Log fixes

* Fix review comments

* Check if logging is initialised in the logger

* Fix plan parser test

* Bring back single config path constructor as deprecated

* Update alica_engine/src/engine/modelmanagement/ModelManager.cpp

Co-authored-by: Maksim Derbasov <ntfs.hard@gmail.com>

* Review comment

---------

Co-authored-by: Maksim Derbasov <ntfs.hard@gmail.com>