Improve type determination in blackboard impl
The type determination method used to default construct a value of the
given type which meant that -
1. an unnecessary object would be created just for determining the type
which could potentially be a costly operation that could have side
effects depending on the type stored
2. the default constructor might not be available, so such types would
not be able to be stored on the blackboard
3. a variant was constructed from the value & the type determined using
the variant index. This is a potentially confusing operation because
the type determination relies on the variant implementation.
To overcome this, the type is determined at compile time following
clear steps to determine supported type conversions:
- ints are promoted to int64 (already existed)
- uints are promoted to uint64 (already existed)
- floats are promoted to double (indirectly supported earlier, but
explicitly added now)
- the promoted type is then checked to see if it is present in the
variant types at compile time
- if not found, the type is treated as std::any