Remove eventDriven & deferring of behaviours
eventDriven is removed because:
- there is no use case for this functionality so far
- external code cannot possibly hold a reference to a behaviour because
there is no such external API & even if such a reference is obtained
it can be invalidated at anytime because the behaviour can be destroyed
at any time & therefore triggering the  behaviour in the current way
will likely lead to crashes
- the functionality of eventDriven can be replicated by having
a state machine with 2 states (A & B) & a transition. The transition
will be true when the trigger condition is true & the code that needs to
be executed (i.e. the code that was in run() method earlier) will now
be in the init method of B

deferring is removed because:
- there is no use case for this functionality so far
- it is not currently implemented
- there is no easy way to implement it either
- same functionality can be had by having an additional state that
waits for the deferring period before transitioning to the state with
the attached behaviour