Logging fixes (#496)
* Logging fixes

- Make AlicaDefaultLogger thread safe since it is the responsibility of
the logging implementation

- Set the logger immediately on construction of the context so that it
can be used to log messages during construction & before init() is
called. This solves some obsure bugs where the logger was used before
it was initialized in the context's set api's eg. setTraceFactory or the
trace factory'sconstructor etc.

- Provides the ability to set the logger multiple times before the
context is initialized which is consistent with other set api's

- Destroy the logger when the context is destroyed

- Add tests for custom logging & fix the test for default logging

* Fix compile error in supplementary tests