Fix crash: dll access after its destruction (#514)
The dynamic creators maintain a reference to the currently loaded
symbol, therefore dlls that were loaded prior would no longer be loaded
& accessing such dlls (typically in the destructor of the beh/plan etc)
would cause undefined behaviour. Specifically this lead to a crash in
the destructor of the GoTo behaviour in turtlesim. Such crashes are
hard to debug & can occur in more ways due to this issue.
As a fix, the creator for each symbol is stored for the lifetime of
the dynamic creator object itself which is basically the entire
duration of the engine. Consequently the symbol & dll is loaded
just once potentially improving performance.