Add ability to set keys on the global blackboard
The master plan can specify if its keys are meant for the global
blackboard using the inherit_blackboard flag in the .plc file.
If true, the keys are populated to the global blackboard & the master
plan simply inherits the global blackboard as its own. If false,
the behaviour remains unchanged.