Fix precision error in ros clock sleep
ros clock used AlicaTime::inSeconds() to convert
the duration of the sleep to seconds. However, this
works out to 0 for any time interval < 1 sec. Since
the sleep time is 33 ms in step engine, this can
cause problems. Futher, this rounding behaviour
is not intuitive for users of the clock in general