No special handling of exceptions thrown from behaviours/plans
The current exception handling in both behaviours & plans (for user
implemented code) traces the exception, logs it & then rethrows the
exception. However, because the exception is caught & then rethrown
the backtrace is lost making debugging impossible. It is better to get
the backtrace than have a log/trace which often contains little to no
information to actually fix the problem that caused the exception in the
first place. Therefore, the exception handling is completely removed.
If such handling is really required it can be implemented in the user
code itself.